# src/user/user.controller.ts
import { Body, Controller, Get, Put, UseGuards, ValidationPipe } from '@nestjs/common';
import { AuthGuard } from '@nestjs/passport';
import { AuthService } from 'src/auth/auth.service';
import { User } from 'src/auth/user.decorator';
import { UserEntity } from 'src/entities/user.entity';
import { UpdateUserDTO } from 'src/models/user.model';

@Controller('user')
export class UserController {
    constructor(private authService: AuthService){}

    @Get()
    @UseGuards(AuthGuard())
    findCurrentUser(@User() {username}:UserEntity){
        return this.authService.currentUser(username);
    }

    @Put()
    @UseGuards(AuthGuard())
    update(@User() {username}: UserEntity, @Body(new ValidationPipe({transform:true, whitelist:true})) data:{user: UpdateUserDTO}){
        return this.authService.updateUser(username, data.user);
    }

}